+++ R/linearIRF.R
@@ -5,8 +5,8 @@
 #'
 #' @inheritParams fitbsSSTVAR
 #' @param stvar an object of class \code{'stvar'} defining a structural or reduced form
-#'   STVAR model. For a reduced form model, the shocks are automatically identified by
-#'   the lower triangular Cholesky decomposition.
+#'   STVAR model. For a reduced form model (that is not readily identified statiscally),
+#'   the shocks are automatically identified by the lower triangular Cholesky decomposition.
 #' @param N a positive integer specifying the horizon how far ahead should the
 #'   linear impulse responses be calculated.
 #' @param regime Based on which regime the linear IRF should be calculated?
@@ -56,8 +56,8 @@
 #'        is then the in a form similar to \code{$point_est}.}
 #'     \item{Other elements:}{contains some of the arguments the \code{linear_IRF} was called with.}
 #'   }
-#' @seealso \code{\link{GIRF}}, \code{\link{GFEVD}}, \code{\link{fitSTVAR}}, \code{\link{STVAR}},
-#'   \code{\link{reorder_B_columns}}, \code{\link{swap_B_signs}}
+#' @seealso \code{\link{GIRF}}, \code{\link{GFEVD}}, \code{\link{hist_decomp}}, \code{\link{cfact_hist}}, \code{\link{cfact_fore}},
+#'  \code{\link{cfact_girf}}, \code{\link{fitSTVAR}}, \code{\link{STVAR}}, \code{\link{reorder_B_columns}}, \code{\link{swap_B_signs}}
 #' @references
 #'  \itemize{
 #'    \item Herwartz H. and Lütkepohl H. 2014. Structural vector autoregressions with Markov switching: